feat: 9주차 미션 완료

+++ 와나/spring/src/main/java/com/umc/spring/dto/requestDto/ReviewRequestDto.java
@@ -7,6 +7,6 @@ public class ReviewRequestDto {
     @Getter
     public static class ReviewCreateDto {
         String content;
-        Double score;
+        Float score;
     }
 }
